# REST APIs Design/openapi.yaml
openapi: 3.0.1
info:
  title: Film Manager
  description: REST APIs for the interaction with the Film Manager service.
  termsOfService: 'http://swagger.io/terms/'
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
  version: 1.0.0
...
servers:
  - url: 'https://todo.swagger.io/v2'
  - url: 'http://todo.swagger.io/v2'
...
paths:
...
  '/api/films/public': 
    get:
      tags:
        - films/public
      summary: Retrieve the public films
      description: >-
        The public films (i.e., the films that are visible for all the users of the service) are retrieved. This operation does not require authentication. A pagination mechanism is implemented to limit the size of messages.
      operationId: getPublicFilms
      parameters:
        - in: query
          name: pageNo
          required: false
          schema:
            type: integer
          description: 'The id of the requested page (if absent, the first page is returned)'
          example: 1
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalPages:
                    type: integer
                  currentPage:
                    type: integer
                  totalItems:
                    type: integer
                  films:
                    type: array
                    items:
                      $ref: '#/components/schemas/Film'
                  next:
                    type: string
                    format: uri
              example:
                totalPages: 1
                currentPage: 1
                totalItems: 1
                films: [{
                  id: 1,
                  title: Your Name,
                  owner: 1,
                  private: false,
                  watchDate: 2021-10-03,
                  rating: 9,
                  favorite: true,
                  self: /api/films/public/1
                }]
        '500':
          description: Internal Server Error
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/Error' 
...
  '/api/films/public/{filmId}/reviews/cooperative/{reviewersId}/drafts':
    post:
      tags:
        - 'films/public/{filmId}/reviews/cooperative/{reviewersId}/drafts'
      security:
        - cookieAuth: []
      summary: Issue a draft of a cooperative film review 
      description: A draft for a cooperative review is created. This operation can only be performed by one of the reviewers. No other draft should have the "open" status for the same review. The review must not be completed. 
      operationId: addCooperativeReviewDraft
      parameters:
        - name: filmId
          in: path
          description: ID of the film
          required: true
          schema:
            type: integer
            format: int64
          example: 1
        - name: reviewersId
          in: path
          description: the list of user IDs to whom the review has been issued
          required: true
          schema:
            type: string
            pattern: ^[0-9]+[,][0-9]([,][0-9]+)*$
          example : 1,2,3
      requestBody:
        description: the new draft of the cooperative film review, including the users to whom they are issued
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                rating:
                  type: integer
                  description: the rating the reviewer has given to the film
                  minimum: 1
                  maximum: 10
                review:
                  type: string
                  description: a textual description of the review
                  maxLength: 1000
              required:
                - rating
                - review
            example:
              rating: 10
              review: Davvero emozionante
        required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  draft:
                    $ref: '#/components/schemas/Draft'
                  agreement:
                    $ref: '#/components/schemas/Agreement'
                required:
                  - draft
              example:
                draft:
                  film: 9
                  reviewersId: [1,2,3]
                  authorId: 1
                  draftId: 1
                  rating: 4
                  review: Sooo boring..
                  closed: false
                  self: /api/films/public/9/reviews/cooperative/-1-2-3-/drafts/1
                agreement:
                  film: 9
                  reviewersId: [1,2,3]
                  authorId: 1
                  draftId: 1
                  agreement: true
                  self: /api/films/public/9/reviews/cooperative/-1-2-3-/drafts/1/agreements/1}
        '400':
          description: Bad Request 
          content: 
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/Error'
                  - $ref: '#/components/schemas/JsonSchemaValidationError'
              example:
                errors: [{
                  param: 'Server',
                  msg: 'Wrong format of filmId path parameter'
                }]
        '401':
          description: Unauthorized
          content: 
            application/json:
              schema:
                type: object
                properties: 
                  error:
                    type: string
                    enum: ['Not authorized']
                additionalProperties: false
                required: 
                  - error
              example:
                error: 'Not authorized'
        '403':
          description: Forbidden (ex. the author ofthe draft is not a reviewer)
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                errors: [{
                  param: 'Server',
                  msg: 'The author of the draft is not a reviewer'
                }]
        '404':
          description: Not Found
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                errors: [{
                  param: 'Server',
                  msg: 'The film with ID filmID does not exist'
                }]
        '409':
          description: Conflict (e.g., the review has already been completed)
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                errors: [{
                  param: 'Server',
                  msg: 'The review has already been completed'
                }]
        '500':
          description: Internal Server Error
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      tags:
        - 'films/public/{filmId}/reviews/cooperative/{reviewersId}/drafts'
      summary: Retrieve the drafts of a cooperative review that has been issued/completed for a film 
      description: >-
        All the drafts of a cooperative review that has been issued/completed for the film with ID filmId are retrieved. A pagination mechanism is implemented to limit the size of messages.
        This operation does not require authentication. If using the query parameter reviewersID, i can return all the cooperative reviews in which the 
        users of the reviewersID are present.
      operationId: getCooperativeReviewDrafts
      parameters:
        - name: filmId
          in: path
          description: ID of the film whose reviews must be retrieved
          required: true
          schema:
            type: integer
            format: int64
          example: 1
        - in: query
          name: pageNo
          required: false
          schema:
            type: integer
          description: ID of the requested page (if absent, the first page is returned)'
          example: 1
        - name: reviewersId
          in: path
          description: the list of user IDs to whom the review has been issued
          required: true
          schema:
            type: string
            pattern: ^[0-9]+[,][0-9]([,][0-9]+)*$
          example : 1,2,3
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalPages:
                    type: integer
                  currentPage:
                    type: integer
                  totalItems:
                    type: integer
                  reviews:
                    type: array
                    items:
                      $ref: '#/components/schemas/Draft'
                  next:
                    type: string
                    format: uri
              example:
                totalPages: 1
                currentPage: 1
                totalItems: 1
                drafts: [{
                  film: 9,
                  reviewersId: [1,2,3],
                  authorId: 1 ,
                  draftId: 1 ,
                  rating: 4 ,
                  review: 'Sooo boring..',
                  closed: false ,
                  self: /api/films/public/9/reviews/cooperative/-1-2-3-/drafts/1          
                }]
        '400':
          description: Bad Request 
          content: 
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/Error'
                  - $ref: '#/components/schemas/JsonSchemaValidationError'
              example:
                errors: [{
                  param: 'Server',
                  msg: 'Wrong format of filmId path parameter'
                }]
        '404':
          description: Not Found
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                errors: [{
                  param: 'Server',
                  msg: 'The film with ID filmID does not exist'
                }]
        '500':
          description: Internal Server Error
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    Film:
      description: schema for Film data structures
      type: object
      additionalProperties: false
      properties:
        $schema:
          type: string
          description: path to the JSON Schema to use for the validation of the JSON data structure
            describing the film
        id:
          type: integer
          description: unique identifier for the film data structure
        title:
          type: string
          description: textual title of the film
        owner:
          type: integer
          description: unique identifier of the the film data structure owner
        private:
          type: boolean
          description: true if only the owner of the film data structure can access it, false otherwise
          default: 'true'
        watchDate:
          type: string
          format: date
          description: the date when the film was watched by the owner
        rating:
          type: integer
          description: the rating the owner has given to the film
          minimum: 1
          maximum: 10
        favorite:
          type: boolean
          description: 'true if the film is among the favourite
            ones of the owner, false otherwise '
          default: 'false'
        self: 
          type: string
          format: uri
      required:
      - title
      - owner
      - private
...
    Draft:
      description: schema for Draft data structures
      type: object
      additionalProperties: false
      properties:
        $schema:
          type: string
          description: path to the JSON Schema to use for the validation of the JSON data structure
            describing the draft
        draftId:
          type: integer
          description: unique identifier of the draft of a review invitation 
        filmId:
          type: integer
          description: unique identifier of the film for which a review invitation has been
            issued
        reviewersId:
          type: array
          items:
            type: integer
          description: an array of unique identifiers of the users who have received the review invitation
        rating:
          type: integer
          description: the rating the reviewer has given to the film
          minimum: 1
          maximum: 10
        review:
          type: string
          description: a textual description of the review
          maxLength: 1000
        authorID:
          type: integer
          description: unique identifier of the author of a review's draft
        closed:
          type: boolean
          description: to specify if the review draft has ben closed or not
        self: 
          type: string
          format: uri
      required:
      - draftId
      - filmId
      - reviewersId
      - authorId
      - closed
      - rating
      - review
    Agreement:
      description: schema for Agreement data structures
      type: object
      additionalProperties: false
      properties:
        $schema:
          type: string
          description: path to the JSON Schema to use for the validation of the JSON data structure
            describing the draft
        draftId:
          type: integer
          description: unique identifier of the draft of a review invitation 
        filmId:
          type: integer
          description: unique identifier of the film for which a review invitation has been
            issued
        reviewersId:
          type: array
          items:
            type: integer
          description: an array of unique identifiers of the users who have received the review invitation
        authorId:
          type: integer
          description: unique identifier of the author of a review's draft's agreement
        agreement:
          type: boolean
          description: to specify if the review draft has ben approved or not
        reason:
          type: string
          description: a textual description of the reason why a user agrees/disagrees with a review draft
          maxLength: 1000
        self: 
          type: string
          format: uri
      required:
      - draftId
      - filmId
      - reviewersId
      - authorId
      - agreement
    Error:
      description: schema for Error data structures
      type: object
      properties:
        errors:
          type: array
          minItems: 1
          maxItems: 1
          items:
            type: object
            properties:
              param:
                type: string
                enum: ['Server']
              msg:
                description: 'Explanation of the error'
                type: string
            additionalProperties: false
            required:
              - param
              - msg
      additionalProperties: false
      required:
      - errors
    JsonSchemaValidationError:
      description: schema for JsonSchemaValidationError data structures
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
          enum: ['JsonSchemaValidationError']
        validationErrors:
          type: object
          properties:
            body:
              type: array
              items:
                type: object
  securitySchemes:
    cookieAuth:
      description: JWT to identify the authenticated user.
      type: apiKey
      in: cookie
      name: connect.sid